Change vehicle counting and tracking mechanism
The vehicle counting is performed on the entire screen as opposed to a line-based checkpoint method.
For each vehicle in the current frame, the shortest distance between this coordinate and the vehicles detected in the last 10 frames is calculated. If the distance is smaller than the length and width of the vehicle, then count does not increment. By doing so, the vehicles can be tracked even if they are not detected in a few frames.